Fixing the Routes directory structure
